Moves Env file to config folder
