--- src/Env.js
@@ -1,6 +1,0 @@
-const Env = {
-    serverAddress: "8563c8fb.ngrok.io",
-    serverPort: "80"
-}
-
-export default Env;
