+++ src/config/Env.js
@@ -1,0 +1,6 @@
+const Env = {
+    serverAddress: "localhost",
+    serverPort: "8081"
+}
+
+export default Env;
